\name{main}
\alias{main}
\title{Try out stuff on Travis}
\usage{
main()
}
\value{
Constantly subject to change depending on what I want to see
}
\description{
Whatever this package ends up doing I'll only be able to see the result through travis  Whatever is returned by this function will hopefully be viewable in the travis output
}
\examples{
# will be run from travis
main()
}
